Added Functionality to Registration
Added Functionality to allow staff to be created in registration form, this is done using an optional field, if left blank user will be created as before. If contents are present they are checked for correct code. If code is incorrect, user is returned to the Login error screen with no further actions made. Otherwise if code is correct, Game Master user will be created